+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fix false positive in `RSpec/Pending`, where it would mark the default block `it` as an offense. ([@bquorning])
 - Fix issue when `Style/ContextWording` is configured with a Prefix being interpreted as a boolean, like `on`. ([@sakuro])
 - Add new `RSpec/IncludeExamples` cop to enforce using `it_behaves_like` over `include_examples`. ([@dvandersluis])
+- Change `RSpec/ScatteredSetup` to allow `around` hooks to be scattered. ([@ydah])
 
 ## 3.5.0 (2025-02-16)